requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
python-telegram-bot==13.14
pytz==2023.3
six==1.16.0
tornado==6.1
tzdata==2023.3
tzlocal==5.0.1
emoji==1.7.0